refactor: 예약 승인 api 위치 변경
