+++ src/main/java/roomescape/member/service/AdminService.java
@@ -6,12 +6,15 @@ import roomescape.member.domain.Member;
 import roomescape.member.dto.request.CreateReservationRequest;
 import roomescape.member.dto.response.CreateReservationResponse;
 import roomescape.member.repository.MemberRepository;
+import roomescape.reservation.dto.response.ConfirmReservationResponse;
 import roomescape.reservation.model.Reservation;
-import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservation.model.ReservationTime;
-import roomescape.reservation.repository.ReservationTimeRepository;
 import roomescape.reservation.model.Theme;
+import roomescape.reservation.model.Waiting;
+import roomescape.reservation.repository.ReservationRepository;
+import roomescape.reservation.repository.ReservationTimeRepository;
 import roomescape.reservation.repository.ThemeRepository;
+import roomescape.reservation.repository.WaitingRepository;
 
 @Service
 public class AdminService {
@@ -19,15 +22,18 @@ public class AdminService {
     private final ReservationTimeRepository reservationTimeRepository;
     private final ThemeRepository themeRepository;
     private final ReservationRepository reservationRepository;
+    private final WaitingRepository waitingRepository;
 
     public AdminService(final MemberRepository memberRepository,
                         final ReservationTimeRepository reservationTimeRepository,
                         final ThemeRepository themeRepository,
-                        final ReservationRepository reservationRepository) {
+                        final ReservationRepository reservationRepository,
+                        WaitingRepository waitingRepository) {
         this.memberRepository = memberRepository;
         this.reservationTimeRepository = reservationTimeRepository;
         this.themeRepository = themeRepository;
         this.reservationRepository = reservationRepository;
+        this.waitingRepository = waitingRepository;
     }
 
     public CreateReservationResponse createReservation(final CreateReservationRequest createReservationRequest) {
@@ -53,5 +59,35 @@ public class AdminService {
     private ReservationTime getReservationTime(Long id) {
         return reservationTimeRepository.findById(id)
                 .orElseThrow(() -> new NoSuchElementException("해당하는 시간이 존재하지 않아 예약을 생성할 수 없습니다."));
+    }
+
+    public ConfirmReservationResponse confirmWaiting(Long id) {
+        Waiting waiting = findWaiting(id);
+        checkReservationExists(waiting);
+        checkFirstWaiting(waiting);
+
+        Reservation reservation = new Reservation(waiting.getMember(), waiting.getDate(), waiting.getReservationTime(),
+                waiting.getTheme());
+        waitingRepository.deleteById(id);
+        return ConfirmReservationResponse.from(reservationRepository.save(reservation));
+    }
+
+    private void checkReservationExists(Waiting waiting) {
+        if (reservationRepository.existsByDateAndReservationTimeIdAndThemeId(
+                waiting.getDate(), waiting.getReservationTime().getId(), waiting.getTheme().getId())) {
+            throw new IllegalArgumentException("이미 예약이 존재하여 대기를 예약으로 변경할 수 없습니다.");
+        }
+    }
+
+    private void checkFirstWaiting(Waiting waiting) {
+        if (waitingRepository.existsByDateAndReservationTimeIdAndThemeIdAndIdLessThan(
+                waiting.getDate(), waiting.getReservationTime().getId(), waiting.getTheme().getId(), waiting.getId())) {
+            throw new IllegalArgumentException(waiting.getId() + "보다 앞선 대기가 존재하여 예약으로 변경할 수 없습니다.");
+        }
+    }
+
+    private Waiting findWaiting(Long id) {
+        return waitingRepository.findById(id)
+                .orElseThrow(() -> new NoSuchElementException("식별자 " + id + "에 해당하는 대기가 존재하지 않아 예약으로 변경할 수 없습니다."));
     }
 }
