+++ src/main/java/roomescape/reservation/controller/WaitingController.java
@@ -14,7 +14,6 @@ import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.core.AuthenticationPrincipal;
 import roomescape.auth.domain.AuthInfo;
 import roomescape.reservation.dto.request.CreateWaitingRequest;
-import roomescape.reservation.dto.response.ConfirmReservationResponse;
 import roomescape.reservation.dto.response.CreateWaitingResponse;
 import roomescape.reservation.dto.response.FindWaitingResponse;
 import roomescape.reservation.service.WaitingService;
@@ -39,17 +38,8 @@ public class WaitingController {
                 .body(createWaitingResponse);
     }
 
-    @PostMapping("/{id}/confirmation")
-    public ResponseEntity<ConfirmReservationResponse> confirmWaiting(
-            @AuthenticationPrincipal AuthInfo authInfo, @PathVariable Long id) {
-        
-        ConfirmReservationResponse confirmReservationResponse = waitingService.confirmWaiting(id);
-        return ResponseEntity.created(URI.create("/waitings/" + confirmReservationResponse.id()))
-                .body(confirmReservationResponse);
-    }
-
     @GetMapping
-    public ResponseEntity<List<FindWaitingResponse>> getReservations() {
+    public ResponseEntity<List<FindWaitingResponse>> getWaitings() {
         return ResponseEntity.ok(waitingService.getWaitings());
     }
