+++ src/main/resources/static/js/waiting.js
@@ -57,7 +57,7 @@ function approve(event) {
     TODO: [4단계] 예약 대기 목록 관리 기능
           예약 대기 승인 API 호출
      */
-    const endpoint = '/waitings/' + id + '/confirmation'
+    const endpoint = '/admin/waitings/' + id + '/confirmation'
     return fetch(endpoint, {
         method: 'POST'
     }).then(response => {
